examples/flowers/requirements.txt
google-api-python-client==1.6.4
numpy==1.14.0
apache-beam[gcp]==2.4.0
tensorflow==1.8.0
tensorflow-transform==0.6.0
Pillow==5.0.0
six==1.10.0